requirements.txt
beautifulsoup4>=4.12.0
mysql-connector-python>=8.0.33
requests>=2.31.0
pandas>=2.0.0
streamlit>=1.28.0
